+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
-## 05/07/2020 - v0.6.4.2
+## 06/08/2020 - v0.6.4.3
+
+* (Add) Pixel Editor - Supports and Drain holes: AntiAliasing
+* (Add) Pixel Editor - Drawing: Line type and defaults to AntiAliasing
+* (Add) Pixel Editor - Drawing: Line thickness to allow hollow shapes
+* (Add) Pixel Editor - Drawing: Layer depth, to add pixels at multiple layers at once
+* (Add) Pixel Editor: Text writing
+
+## 05/08/2020 - v0.6.4.2
 
 * (Add) Hold "ALT" key when double clicking over items to invert AutoZoom setting, prevent or do zoom in issues or pixels, this will behave as !AutoZoom as long key is held
 * (Improvement) Partial island update speed, huge boost performance over large files
 
-## 04/07/2020 - v0.6.4.1
+## 04/08/2020 - v0.6.4.1
 
 * (Add) Partial update islands from current working layer and next layer when using pixel editor or island remove
 * (Add) Setting: To enable or disable partial update islands